refactor(ffmpeg): extract common FFmpeg process logic into BaseProcess
- Create abstract BaseProcess class with common FFmpeg process management:
  - Handles process spawning, stdout/stderr capture, and error handling
  - Provides runProcessing method for executing FFmpeg commands
  - Includes kill and isRunning methods
- Refactor ProcessManager to extend BaseProcess, reducing duplication
- Refactor TwoPassProcessManager to extend BaseProcess and use runProcessing
- Remove duplicate spawning and event handling code from both classes
