+++ src/main/services/ffmpeg/twoPassProcessManager.ts
@@ -15,8 +15,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-import { type ChildProcessWithoutNullStreams, spawn } from "node:child_process";
-import { EventEmitter } from "node:events";
 import fs from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
@@ -24,7 +22,7 @@ import {
 	buildLoudnormFirstPassOptions,
 	buildLoudnormSecondPassOptions,
 } from "@main/lib/ffmpeg/utils/buildLoudnormOptions";
-import ffmpegPath from "ffmpeg-static";
+import { BaseProcess } from "./baseProcess";
 
 export type LoudnormTwoPassOptions = {
 	input: string;
@@ -59,20 +57,8 @@ type Stats = {
 	measured_offset: number;
 };
 
-export class TwoPassProcessManager extends EventEmitter {
-	private proc: ChildProcessWithoutNullStreams | null = null;
+export class TwoPassProcessManager extends BaseProcess {
 	private tempDir: string | null = null;
-	private ffmpeg: string;
-
-	constructor() {
-		super();
-		if (!ffmpegPath) {
-			throw new Error(
-				"FFmpeg binary not found. Please ensure ffmpeg-static is installed correctly.",
-			);
-		}
-		this.ffmpeg = ffmpegPath;
-	}
 
 	async run(options: LoudnormTwoPassOptions): Promise<void> {
 		const { input, output, globalSettings, filterOptions, signal } = options;
@@ -97,7 +83,7 @@ export class TwoPassProcessManager extends EventEmitter {
 		];
 
 		// run first pass
-		await this.runPass(firstPassArgs);
+		await this.runProcessing(firstPassArgs);
 
 		// read and parse statistics
 		let stats: LoudnormStats;
@@ -132,91 +118,14 @@ export class TwoPassProcessManager extends EventEmitter {
 		];
 
 		// run second pass
-		await this.runPass(secondPassArgs);
+		await this.runProcessing(secondPassArgs);
 
 		// cleanup temporary files
 		await this.cleanup();
 	}
 
-	private runPass(args: string[]): Promise<void> {
-		let resolved = false;
-		let stderr = "";
-		const MAX_STDERR_BUFFER = 1024 * 1024;
-
-		if (!ffmpegPath) {
-			return Promise.reject(new Error("FFmpeg binary not found."));
-		}
-
-		return new Promise((resolve, reject) => {
-			try {
-				this.proc = spawn(this.ffmpeg, args, {
-					windowsHide: true,
-					stdio: ["pipe", "pipe", "pipe"],
-				});
-			} catch (error) {
-				reject(
-					new Error(
-						`Failed to spawn FFmpeg: ${error instanceof Error ? error.message : String(error)}`,
-					),
-				);
-			}
-
-			const stdoutHandler = (data: Buffer) =>
-				this.emit("stdout", data.toString());
-			const stderrHandler = (data: Buffer) => {
-				const chunk = data.toString();
-				stderr += chunk;
-				if (stderr.length > MAX_STDERR_BUFFER) {
-					stderr = stderr.slice(-MAX_STDERR_BUFFER / 2);
-				}
-				this.emit("stderr", chunk);
-			};
-
-			this.proc?.stdout.on("data", stdoutHandler);
-			this.proc?.stderr.on("data", stderrHandler);
-
-			const cleanup = () => {
-				if (this.proc) {
-					this.proc.stdout.off("data", stdoutHandler);
-					this.proc.stderr.off("data", stderrHandler);
-					this.proc.removeAllListeners();
-				}
-			};
-
-			const onError = (err: Error) => {
-				if (resolved) return;
-				resolved = true;
-				cleanup();
-				reject(new Error(`FFmpeg pass error: ${err.message}`));
-			};
-
-			const onClose = (code: number | null, signal: NodeJS.Signals) => {
-				if (resolved) return;
-				resolved = true;
-				cleanup();
-
-				if (code === 0 || code === 255) {
-					resolve();
-				} else if (signal) {
-					reject(
-						new Error(`FFmpeg process was terminated by signal: ${signal}`),
-					);
-				} else {
-					reject(
-						new Error(`FFmpeg process exited with code ${code}\n${stderr}`),
-					);
-				}
-			};
-
-			this.proc?.once("error", onError);
-			this.proc?.once("close", onClose);
-		});
-	}
-
 	async kill(signal: NodeJS.Signals = "SIGINT"): Promise<void> {
-		if (this.proc) {
-			this.proc.kill(signal);
-		}
+		super.kill(signal);
 		await this.cleanup();
 	}
 
@@ -225,14 +134,9 @@ export class TwoPassProcessManager extends EventEmitter {
 			try {
 				await fs.rm(this.tempDir, { recursive: true, force: true });
 			} catch {
-				// Non-critical, log if needed but don't throw
 			} finally {
 				this.tempDir = null;
 			}
 		}
-	}
-
-	isRunning(): boolean {
-		return this.proc !== null;
 	}
 }
